Upate test files
