+++ src/test/saveandrendertask.test.js
@@ -1,15 +1,23 @@
-import {
-  saveAndRender,
-  render,
-  lists,
-  selectedListId,
-  createList,
-} from '../modules/saveandrender';
+import { saveAndRender, createList } from '../modules/saveandrender'
 
-jest.mock('../modules/saveandrender');
+describe('Create a new list', () => {
+  let id = Date.now().toString()
+  let listName = 'New List'
+  let newList = createList(listName, id)
 
+  test('Successfully create a new list', () => {
+    expect(typeof newList).toBe('object')
+  })
+
+  test('The id of a list should be truthy', () => {
+    expect(newList.id).not.toBeNull()
+  })
+
+  test('Should return the value of the list name', () => {
+    expect(newList.name).toEqual('New List')
+  })
+})
 
 test('Check the presence of task list', () => {
-  expect(saveAndRender()).not.toBeNull();
-});
-
+  expect(saveAndRender()).not.toBeNull()
+})
